Bootloader automatically boots into main app
When the device is reset, the bootloader will check the validity of the
main application (app1). If it is valid, it will switch to app1. The
user can override this by holding SW2 and the reset button at the same
time, releasing the reset button and continuing to hold SW2 down for at
least .5 seconds. Then, upon releasing SW2, the bootloader will begin
executing and not switch to app1.