Add advertising timeout and hibernate ability
The DFU device will spend 30 seconds in fast advertising mode. After
which, if no connection is made, it will spend 150 seconds in slow
advertising mode. After which, if no connection is made, it will attempt
to validate the main application (app1). If app1 is valid, it will
switch to it. If app1 is not valid, the device will enter hibernate
mode. The user must press SW2 to wake the device from hibernate.