Add IAS functionality
Adds the Immediate Alert Service (IAS) to the DFU BLE Server profile.
This allows the user to connect to the BLE DFU device and write to the
Alert Level characteristic. An alert level of 1 or 2 will tell the DFU
device to switch to the main application (app1) if it is valid. Any
other alert value will do nothing.